+++ src/test/kotlin/io/github/armanayvazyan/propsy/PropsyTableModelTest.kt
@@ -8,7 +8,7 @@ class PropsyTableModelTest : BasePlatformTestCase() {
     fun testModelExposesKeyValueColumns() {
         val file = myFixture.configureByText("m.properties", "foo=1\nbar=2\n") as PropertiesFile
         val model = PropsyTableModel(project)
-        model.load(file)
+        model.load(ResolvedFile(PropertiesBackend(), file))
         assertEquals(2, model.rowCount)
         assertEquals(2, model.columnCount)
         assertEquals("foo", model.getValueAt(0, 0))
@@ -18,7 +18,7 @@ class PropsyTableModelTest : BasePlatformTestCase() {
     fun testSetValueAtWritesThrough() {
         val file = myFixture.configureByText("m.properties", "foo=1\n") as PropertiesFile
         val model = PropsyTableModel(project)
-        model.load(file)
+        model.load(ResolvedFile(PropertiesBackend(), file))
         model.setValueAt("99", 0, 1)
         assertEquals("99", model.getValueAt(0, 1))
         assertEquals("foo=99\n", file.containingFile.text)
@@ -27,7 +27,7 @@ class PropsyTableModelTest : BasePlatformTestCase() {
     fun testSetKeyAtRenames() {
         val file = myFixture.configureByText("m.properties", "foo=1\n") as PropertiesFile
         val model = PropsyTableModel(project)
-        model.load(file)
+        model.load(ResolvedFile(PropertiesBackend(), file))
         model.setValueAt("renamed", 0, 0)
         assertEquals("renamed", model.getValueAt(0, 0))
         assertEquals("renamed=1\n", file.containingFile.text)